+++ private/enterprise/causal-clustering/src/main/java/com/neo4j/causalclustering/core/state/machines/tx/ByteArrayReplicatedTransaction.java
@@ -28,7 +28,7 @@ public class ByteArrayReplicatedTransaction extends ReplicatedTransaction
     @Override
     public OptionalLong size()
     {
-        return OptionalLong.of( (long) txBytes.length );
+        return OptionalLong.of( txBytes.length );
     }
 
     @Override
